+++ types/ThreatIntel.ts
@@ -81,109 +81,3 @@ export interface ThreatIntelIocData {
   specVersion: string;
   version: number;
 }
-
-export const dummyIoCDetails: ThreatIntelIocData = {
-  id: 'indicator--8e2e2d2b-17d4-4cbf-938f-98ee46b3cd3f',
-  name: 'my-bad-ip',
-  type: ThreatIntelIocType.IPAddress,
-  value: '192.0.2.1',
-  severity: 'High',
-  created: 1718761171,
-  modified: 1718761171,
-  description: 'Random IP address',
-  labels: [],
-  feedId: 'random-feed-id',
-  specVersion: '',
-  version: 1,
-};
-
-// export interface ThreatIntelSourceItem {
-//   id: string;
-//   version: number;
-//   feedName: string;
-//   description: string;
-//   feedFormat: string;
-//   feedType: FeedType;
-//   createdByUser: string;
-//   createdAt: number;
-//   enabledTime: number;
-//   lastUpdateTime: number;
-//   schedule: string;
-//   state: string;
-//   refreshType: string;
-//   lastRefreshedTime: number;
-//   lastRefreshedUser: string;
-//   isEnabled: boolean;
-//   iocMapStore: Record<string, object>;
-//   iocTypes: string[];
-// }
-
-export const dummySource: ThreatIntelSourceItem = {
-  id: 'hello-world',
-  feedName: 'AlienVault',
-  description: 'Short description for threat intel source',
-  isEnabled: false,
-  iocTypes: ['IP', 'Domain', 'File hash'],
-};
-
-export const dummyLogSource: ThreatIntelLogSource = {
-  name: 'windows*',
-  iocConfigMap: {
-    [ThreatIntelIocType.IPAddress]: {
-      enabled: true,
-      fieldAliases: ['src_ip', 'dst.ip'],
-    },
-    [ThreatIntelIocType.Domain]: {
-      enabled: true,
-      fieldAliases: ['domain'],
-    },
-    [ThreatIntelIocType.FileHash]: {
-      enabled: false,
-      fieldAliases: ['hash'],
-    },
-  },
-};
-
-export const dummyLogSource2: ThreatIntelLogSource = {
-  name: 'cloudtrail*',
-  iocConfigMap: {
-    [ThreatIntelIocType.IPAddress]: {
-      enabled: true,
-      fieldAliases: ['src_ip', 'dst.ip'],
-    },
-    [ThreatIntelIocType.Domain]: {
-      enabled: true,
-      fieldAliases: ['domain'],
-    },
-    [ThreatIntelIocType.FileHash]: {
-      enabled: false,
-      fieldAliases: ['hash'],
-    },
-  },
-};
-
-/**
- * 
- * 
- * 
- * 
- *  private String id;
-    private Long version;
-    private String feedName;
-    private String feedFormat;
-    private FeedType feedType;
-    private String createdByUser;
-    private Instant createdAt;
-
-    //    private Source source; TODO: create Source Object
-    private Instant enabledTime;
-    private Instant lastUpdateTime;
-    private IntervalSchedule schedule;
-    private TIFJobState state;
-    public RefreshType refreshType;
-    public Instant lastRefreshedTime;
-    public String lastRefreshedUser;
-    private Boolean isEnabled;
-    private Map<String, Object> iocMapStore;
-    private List<String> iocTypes;
- */
